フィードバック内容通りに修正
