+++ web/app/Http/Controllers/CartController.php
@@ -9,7 +9,7 @@ use DB;
 
 class CartController extends Controller
 {
-    //
+    //商品をカートに追加
     function addCart(Request $request)
     {        
         $user = Auth::user();
@@ -27,6 +27,7 @@ class CartController extends Controller
         return response()->json(['cart' => $cart]);
     }
 
+    //カートにある商品表示
     function viewCart(Request $request)
     {
         $user = Auth::user();
@@ -42,6 +43,7 @@ class CartController extends Controller
         return response()->json(['carts'=>$carts]);
     }
 
+    //カートにある商品を削除
     function deleteCart(Cart $cart)
     {
         $cart->delete();
