+++ web/app/Http/Controllers/ProductsController.php
@@ -30,7 +30,7 @@ class ProductsController extends Controller
         return response()->json(['status' => 200000]);
     
     }
-    //
+    //商品情報の表示
     public function productList(Request $request)
     {
         $data = $request->all();
@@ -44,7 +44,8 @@ class ProductsController extends Controller
 
         return response()->json(['products' => $products]);
     }
-    //
+
+    //商品画像登録
     public function imageUpload(Request $request)
     {
         
@@ -87,7 +88,8 @@ class ProductsController extends Controller
         
         return response()->json(['image_url' => Storage::disk('public')->url($path)]); 
     }
-    //
+
+    //商品詳細情報
     public function productDetail($id)
     {
         $product = DB::table('products')
@@ -98,7 +100,8 @@ class ProductsController extends Controller
 
         return response()->json(['product' => $product]);
     }
-    //
+
+    //商品情報の変更
     public function productUpdate(Request $request)
     {
         $productInfo = $request['product'];
@@ -115,7 +118,7 @@ class ProductsController extends Controller
         return response()->json(['product' => $product]);
         
     }
-    //
+    //商品情報の変更のための商品情報の表示
     public function productInfo(Request $request, string $productId)
     {
         $product = DB::table('products')
