+++ web/app/Http/Kernel.php
@@ -33,7 +33,7 @@ class Kernel extends HttpKernel
             \Illuminate\Session\Middleware\StartSession::class,
             // \Illuminate\Session\Middleware\AuthenticateSession::class,
             \Illuminate\View\Middleware\ShareErrorsFromSession::class,
-            //\App\Http\Middleware\VerifyCsrfToken::class,
+            //\App\Http\Middleware\VerifyCsrfToken::class,  //TODO: プロジェクトが終わったらコメント削除の予定
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
         ],
