+++ web/resources/js/pages/products/ProductInfo.vue
@@ -9,7 +9,7 @@
         <li>Product Description: {{ product.description }}</li>
         <router-link v-if="isAdmin" class="btn btn-primary" :to="`/product/${product.id}/edit`">Update</router-link>
         <span v-if="!isAdmin">
-          <button @click="goToAccountView()" class="btn btn-primary">Buy</button>
+          <button @click="goToAccountView" class="btn btn-primary">Buy</button>
           <button v-on:click="counter += 1" class="btn btn-danger">Add Cart</button>
           <p> Added Cart Quantity: {{ counter }} </p>
           <p> Total Price($): {{ counter * product.price }} </p> 
@@ -48,6 +48,7 @@
       .catch(error => console.log(error));
     },
     methods:{
+      //カートに追加する処理
       AddCart(){
         axios.post('/api/cart',{
           cart:this.cart,
